refactor: move ChatEntity back to protocol
It doesn't belong to messaging module.
